Mise en place d'un pannel administrateur pour la gestion de sauvegarde et restauration de la bd et fichiers.
