+++ php/include.php
@@ -35,7 +35,7 @@ require_once '../app/controllers/controller_faq.class.php';
 require_once '../app/controllers/controller_polConf.class.php';
 require_once '../app/controllers/controller_pageErreur.class.php';
 require_once '../app/controllers/controller_newsLetter.class.php';
-
+require_once '../app/controllers/controller_admin.class.php';
